reusable timeline configurator
