+++ src/includes/common.tsx
@@ -2,3 +2,4 @@ export const REPL_DEVHUB = "devhub.near";
 export const REPL_INFRASTRUCTURE_COMMITTEE = "infrastructure-committee.near";
 export const REPL_INFRASTRUCTURE_COMMITTEE_CONTRACT = "truedove38.near";
 export const REPL_RPC_URL = "https://rpc.mainnet.near.org";
+export const REPL_NEAR = "near";
